% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_readme.R
\name{criar_readme}
\alias{criar_readme}
\title{Cria arquivo README para cursos da Curso-R}
\usage{
criar_readme()
}
\value{

}
\description{
Cria o arquivo README responsável pelo ghpage de cada turma dos cursos
da Curso-R. Essa função é de uso interno da Curso-R.
}
