% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func1.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Encontrar o pior ano com menor numeros de vitorias de um time na serie historica.}
\usage{
encontrar_pior_ano_time(time)
}
\arguments{
\item{time}{Nome de um time ('Cruzeiro' por padrao)}
}
\value{
Uma tibble com tres colunas
}
\description{
Encontrar o pior ano com menor numeros de vitorias de um time na serie historica.
}
\examples{

encontrar_pior_ano_time('Fortaleza')
}
