\name{inst.pack.cursoR}
\alias{inst.pack.cursoR}
\title{Instala diversos pacotes em um único procedimento}
\usage{
inst.pack.cursoR(pkg)
}
\description{
Instala diversos pacotes em um único procedimento

}

\arguments{
 \item{pkg}{Um vetor de pacotes a serem instalados}
 }
 

\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}


\examples{
inst.pack.cursoR(c("ggplo2", "ExpDes"))
}
