% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instalar_dependencias.R
\name{instalar_dependencias}
\alias{instalar_dependencias}
\title{Atualiza o material de um curso}
\usage{
instalar_dependencias()
}
\description{
Esta função instala todos os pacotes utilizados em um determinado curso.
}
\examples{
\dontrun{
instalar_dependencias()
}

}
