% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respostas.R
\name{sessao1}
\alias{sessao1}
\title{Para conferir as respostas dos exercícios continuos no "Exercicios.Rmd"
Operadores e operações básicas}
\usage{
sessao1(nome = FALSE, a1 = FALSE, b1 = FALSE, c1 = FALSE, a2 = FALSE,
  b2 = FALSE, c2 = FALSE, d2 = FALSE, a3 = FALSE)
}
\description{
Para conferir as respostas dos exercícios continuos no "Exercicios.Rmd"
Operadores e operações básicas
}
