% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respostas.R
\name{sessao2}
\alias{sessao2}
\title{Classe de objetos}
\usage{
sessao2(nome = FALSE, a1 = FALSE, b1 = FALSE, c1 = FALSE, a2 = FALSE,
  b2 = FALSE, a3 = FALSE, b3 = FALSE, c3 = FALSE)
}
\description{
Classe de objetos
}
