\name{summary_ge_stats}
\alias{summary_ge_stats}
\title{Summary a ge_stats object}
\usage{
summary_ge_stats(x, export = FALSE, file.name = "ge_stats", digits = 4)
}

\description{
Summary the  \code{ge_stats} object in two ways. By default, the results are shown in the R console. The results can also be exported to the directory.

}

\arguments{

  \item{x}{The \code{ge_stats} object}
  
  \item{export}{A logical argument. If \code{TRUE|T}, a *.txt file is exported to the working directory}

  \item{file.name}{The name of the file if \code{export = TRUE} }

  \item{digits}{The number of significant digits to be shown.}
  
}

\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}




