\name{supresp_col}
\alias{supresp_col}
\title{Cores personalizadas para gráficos de superfície de resposta}
\usage{
supresp_col(col = c("grey", "black"), n = 100, alpha = 1) 
                       
}
\description{
Função para interpolação de cores. Permite criar um vetor de cores, úteis na confecção de gráficos de superfície de resposta.

}

\arguments{

\item{col}{O vetor de cores a serem interpoladas}

\item{n}{O número de cores a serem geradas}
 
\item{alpha}{A transparência das cores. Varia de 0(totalmente transpararente) a 1 (totalmente opaca).}

\author{Tiago Olivoto \email{tiagoolivoto@gmail.com}}




