% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_material.R
\name{atualizar_material}
\alias{atualizar_material}
\title{Atualiza o material de um curso}
\usage{
atualizar_material(curso)
}
\arguments{
\item{curso}{Uma string indicando o curso que você está fazendo. Veja
as opções abaixo.}
}
\description{
Esta função baixa o repositório do curso no Github e atualiza o material
criando ou substituindo apenas os arquivos novos. Para funcionar
corretamente, você precisa estar com o projeto do curso aberto.
}
\details{
As opções do argumento curso são:
\itemize{
\item r4ds1: R para Ciência de Dados 1
}
}
\examples{
# Com o projeto do curso R para Ciência de Dados 1 aberto, rode:

atualizar_matarial("r4ds1")
}
