% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_material.R
\name{atualizar_material}
\alias{atualizar_material}
\title{Atualiza o material de um curso}
\usage{
atualizar_material()
}
\description{
Esta função baixa o repositório do curso no Github e atualiza o material
criando ou substituindo apenas os arquivos novos. Para funcionar
corretamente, você precisa estar com o projeto do curso aberto.
}
\examples{
\dontrun{
# Com o projeto do seu curso aberto, rode:
atualizar_material()
}

}
